+++ containers/scripts/start_okit_server_win.ps1
@@ -1,4 +1,7 @@
 
+# Copyright (c) 2020, Oracle and/or its affiliates.
+# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
+
 #Start Docker image on windows
 #Assuming that OCI CLI it's configured in default path - oci setup config
 
@@ -6,8 +9,8 @@ Clear-Host
 
 $Scriptfolder =  Split-Path -Path $PSScriptRoot 
 Write-Host $Scriptfolder
-                #Split-Path -Path $PSScriptRoot -Parent
-$Rootfolder= Split-Path -Path $Scriptfolder #(get-item $Scriptfolder ).parent
+               
+$Rootfolder= Split-Path -Path $Scriptfolder 
 Write-Host $Rootfolder
 
 $LogFolder = Join-Path -Path $Rootfolder -ChildPath "log"
@@ -41,9 +44,5 @@ if (-not (Test-Path -LiteralPath $OciCliFolder))
 
 }
 
-#$params = 'run','--rm','-it','-p' ,'443:443/tcp','-p', '80:80/tcp','--hostname', 'okit','--name','okit', '-v', "$($OciCliFolder):/root/.oci:Z",  '-v',"$($OkitwebFolder):/okit/okitweb:Z", '-v' ,"$($VisualizerFolder):/okit/visualiser:Z", '-v',"$($LogFolder):/okit/log:Z" ,'okit:latest'
-#Write-Host  $params
-#& 'docker.exe'  @params
 
-docker run --rm -it -p 443:443/tcp -p 80:80/tcp --hostname okit --name okit -v $ENV:UserProfile\.oci:/root/.oci:Z  -v $Rootfolder\okitweb:/okit/okitweb:Z -v $Rootfolder\visualiser:/okit/visualiser:Z -v $Rootfolder\log:/okit/log:Z okit:latest
-
+& "docker" run --rm -it -p 443:443/tcp -p 80:80/tcp --hostname okit --name okit -v $ENV:UserProfile\.oci:/root/.oci  -v $Rootfolder\okitweb:/okit/okitweb -v $Rootfolder\visualiser:/okit/visualiser -v $Rootfolder\log:/okit/log okit:latest