+++ okitweb/static/okit/js/okit_designer_properties.js
@@ -41,6 +41,9 @@ function setDescendantProp(obj, desc, value) {
     return obj[arr[0]] = value;
 }
 
+let error_propeties = [];
+let warning_propeties = [];
+
 function loadPropertiesSheet(json_element) {
     console.groupCollapsed('Loading Properties');
     $.each(json_element, function(key, val) {
@@ -80,9 +83,9 @@ function loadPropertiesSheet(json_element) {
             console.info(key + ' is label.');
             if (key.endsWith('_id')) {
                 // Get Artifact Associated With Id
-                let artifact_type = key.substr(0, (key.length - 3));
-                console.info('Label : Artifact Type ' + titleCase(artifact_type) + ' - ' + key);
-                $(jqId(key)).html(okitJson['get' + titleCase(artifact_type)](val).display_name);
+                let artefact_type = key.substr(0, (key.length - 3));
+                console.info('Label : Artifact Type ' + titleCase(artefact_type) + ' - ' + key);
+                $(jqId(key)).html(okitJsonView['get' + titleCase(artefact_type)](val).display_name);
             } else {
                 $(jqId(key)).html(val);
             }
@@ -137,7 +140,8 @@ function loadPropertiesSheet(json_element) {
                     .attr('type', 'button')
                     .text('X');
                 button.on('click', function () {
-                    delete json_element.defined_tags[key];
+                    delete json_element.defined_tags[namespace][key];
+                    if (Object.keys(json_element.defined_tags[namespace]).length === 0) {delete json_element.defined_tags[namespace];}
                     loadPropertiesSheet(json_element);
                     d3.event.stopPropagation();
                 });
@@ -152,6 +156,16 @@ function loadPropertiesSheet(json_element) {
     if (okitSettings.is_optional_expanded) {
         d3.select(d3Id("optional_properties")).attr("open", "open");
     }
+    // Check for Errors & Warnings
+    for (let property_name of error_propeties) {
+        $(jqId(property_name)).addClass('okit-error');
+        $(jqId(property_name)).focus();
+    }
+    error_propeties = [];
+    for (let property_name of warning_propeties) {
+        $(jqId(property_name)).addClass('okit-warning');
+    }
+    warning_propeties = [];
     console.groupEnd();
 }